requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
pydantic-settings
python-dotenv
python-multipart
streamlit

# Para pruebas
pytest
httpx
